+++ README.md
@@ -83,8 +83,16 @@ _✨ 使用人工智能来一起画画吧! ✨_
 ### 指令表
 | 指令 | 需要@ | 范围 | 说明 |
 |:-----:|:----:|:----:|:----:|
-| 绘画/画画/画图/作图/绘图/约稿 |  否 | 群聊/私聊 | 使用描述性文本生成图画, 可用参数见[文本生成参数](###文本生成参数) |
+| 绘画/画画/画图/作图/绘图/约稿 |  否 | 群聊/私聊 | 使用描述性文本生成图画, 可用参数见[文本生成参数](###文本生成参数), 管理参数见[绘图管理参数](###绘图管理参数) |
 | 以图绘图/以图生图/以图制图 | 否 | 群聊/私聊 | 在基准图像上使用描述性文本生成图画, 支持回复图片消息使用,<br>可用参数见[图像生成参数](###图像生成参数) |
+
+使用示例：
+
+    /绘图 描述文本 -p l --scale 12
+
+**注意**
+
+默认情况下, 您应该在指令前加上命令前缀, 通常是 /
 
 ### 文本生成参数
 | 参数名 | 简写 | 全写 | 说明 |
@@ -100,10 +108,19 @@ _✨ 使用人工智能来一起画画吧! ✨_
 |:-----:|:----:|:----:|:----:|
 | strength |  -e | --strength | 允许 AI 改变图像的构成, 降低该值会产生更接近原始图像的效果 |
 
+### 绘图管理参数
+
+| 参数名  | 说明 |
+|:-----:|:----:|
+| 查看白名单 | 查看白名单模式下允许的群组 |
+| 查看黑名单 | 查看黑名单模式下禁止的群组 |
+| 添加白名单 + 群号 | 将群组添加到白名单中, 群号以逗号分隔 |
+| 添加黑名单 + 群号 | 将群组添加到黑名单中, 群号以逗号分隔 |
+| 删除白名单 + 群号 | 将群组从白名单中移除, 群号以逗号分隔 |
+| 删除黑名单 + 群号 | 将群组从黑名单中移除, 群号以逗号分隔 |
+| 切换白名单 | 切换到白名单模式, 只有白名单中的群组才允许使用 |
+| 切换黑名单 | 切换到黑名单模式, 只有黑名单中的群组才禁止使用 |
+
 使用示例：
 
-    /绘图 描述文本 -p l --scale 12
-
-**注意**
-
-默认情况下, 您应该在指令前加上命令前缀, 通常是 /
+    /绘图添加黑名单 123456
